refactor(api): move apply patch method
